# docs/swagger.yaml
basePath: /
definitions:
...
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      request_id:
        type: string
      trace_id:
        type: string
    type: object
info:
  contact: {}
  description: 企业级 Gin 脚手架示例 API
  title: Gin Scaffold API
  version: "1.0"
paths:
  /api/v1/admin/audit-logs:
    get:
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 分页大小（<=200）
        in: query
        name: page_size
        type: integer
      - description: 用户 ID
        in: query
        name: user_id
        type: integer
      - description: 动作：POST|PUT|PATCH|DELETE
        in: query
        name: action
        type: string
      - description: HTTP 状态码
        in: query
        name: status
        type: integer
      - description: 按路径模糊匹配
        in: query
        name: path
        type: string
      - description: 按请求 ID 精确匹配
        in: query
        name: request_id
        type: string
      - description: 开始时间（RFC3339）
        in: query
        name: from
        type: string
      - description: 结束时间（RFC3339）
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: 审计日志查询（后台）
      tags:
      - admin-ops
  /api/v1/admin/audit-logs/export/tasks:
    post:
      parameters:
      - description: 用户 ID
        in: query
        name: user_id
        type: integer
      - description: 动作：POST|PUT|PATCH|DELETE
        in: query
        name: action
        type: string
      - description: HTTP 状态码
        in: query
        name: status
        type: integer
      - description: 按路径模糊匹配
        in: query
        name: path
        type: string
      - description: 按请求 ID 精确匹配
        in: query
        name: request_id
        type: string
      - description: 开始时间（RFC3339）；为空时默认 now-export_default_days
        in: query
        name: from
        type: string
      - description: 结束时间（RFC3339）；为空时默认 now
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: 创建审计日志异步导出任务（后台）
      tags:
      - admin-ops
  /api/v1/admin/audit-logs/export/tasks/{task_id}:
    get:
      parameters:
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Body'
      summary: 查询审计日志异步导出任务状态（后台）
      tags:
      - admin-ops
  /api/v1/admin/audit-logs/export/tasks/{task_id}/download:
    get:
      parameters:
      - description: 任务ID
        in: path
        name: task_id
        required: true
        type: string
      produces:
      - text/csv
      responses:
        "200":
          description: OK
          schema:
            type: file
      summary: 下载审计日志异步导出结果（后台）
      tags:
      - admin-ops
...
swagger: "2.0"